Module reshuffle for 1.5-D and API reshuffle for 2-D fixed boundary equilibrium (#2087)
* change default p_order to 2

* add blank file

* add docstring and copyright

* copy across class

* remove old class

* move test file

* rename test file

* remove unused variables

* flake8

* add test_generated_data

* remove unused stuff

* put back not useless stuff

* move FE utilities to the best place

* move b_coil_axis

* b_coil_axis docstring

* docstring uncertainty

* 0.5 statt /2

* move example

* add magnetostatics example

* remove from other examples

* slight docstring mod

* fix examples rst

* add coupling file

* move procedure

* move equilibrium tests

* move test data

* add disclaimer in docstring

* move example and rename bad file

* reshuffle examples.rst

* eudemo imports

* fix more imports

* add fixed boundary example

* update examples rst

* move create_mesh to fem gs utilities

* working example

* a lot of API fixes now its separated!

* change curr target to I_p

* dodge reassign

* fix typing

* flake8

* change default p_order

* start on FemGS API change

* still working without tests :)

* remove boundaries from mesh

* remove BCs from solve

* pause for test

* add test file

* dissolve process profiles and move define g to solve

* flake8 and working tests

* add plot test

* better example

* better better example

* adjusted signature and docstring of Bz_coil_axis

* Bz_coil_axis docstring

* Bz_coil_axis docstring

* add TDO and note in example

* remove opt from func

* remove supporting funcs and flkae8

* ammend docstring

---------
