+++ documentation/source/examples.rst
@@ -22,8 +22,7 @@ Equilibria Examples
 
     examples/equilibria/single_null
     examples/equilibria/eudemo_2017
-    examples/equilibria/fem_fixed_boundary/equilibrium_plasmod_example
-    examples/equilibria/fem_fixed_boundary/fem_magnetostatic_2D_single_coil
+    examples/equilibria/fixed_boundary
 
 Geometry Examples
 ^^^^^^^^^^^^^^^^^
@@ -34,6 +33,15 @@ Geometry Examples
     examples/geometry/optimisation_tutorial
     examples/geometry/plane_placement_tutorial
 
+Magnetostatics Examples
+^^^^^^^^^^^^^^^^^^^^^^^
+
+.. toctree::
+    :maxdepth: 1
+
+    examples/magnetostatics/helmholtz_example
+    examples/magnetostatics/finite_element_2D_single_coil
+
 External Code Examples
 ^^^^^^^^^^^^^^^^^^^^^^
 
@@ -42,6 +50,7 @@ External Code Examples
 
     examples/codes/external_code
     examples/codes/run_plasmod_example
+    examples/codes/equilibrium_plasmod_example
     examples/codes/solver_example
 
 Other Examples
@@ -53,7 +62,6 @@ Other Examples
     examples/base/units
     examples/balance_of_plant/steady_state_example
     examples/fuel_cycle/EUDEMO_fuelcycle
-    examples/magnetostatics/helmholtz_example
     examples/materials/material_definitions
     examples/mesh/mesh_tutorial
     examples/plasma_physics/fusion_reactions
