+++ tests/magnetostatics/test_generated_data/.gitignore
@@ -1,0 +1,2 @@
+*
+!.gitignore
